have no idea how, but partly works

+++ module #3. REST API Advanced/rest-advanced/repository/src/main/java/com/epam/esm/repository/impl/hibernate/TagHibernateRepository.java
@@ -3,6 +3,7 @@ package com.epam.esm.repository.impl.hibernate;
 import com.epam.esm.core.model.Tag;
 import com.epam.esm.repository.TagRepository;
 import jakarta.persistence.EntityManager;
+import jakarta.persistence.NoResultException;
 import jakarta.persistence.PersistenceContext;
 import jakarta.persistence.TypedQuery;
 import lombok.Getter;
@@ -19,63 +20,89 @@ import java.util.Optional;
 @RequiredArgsConstructor
 @Getter
 public class TagHibernateRepository implements TagRepository {
+    private static final String FIND_SINGLE_BY_NAME = "SELECT t FROM Tag t WHERE t.name= :name";
+    private static final String FIND_ALL_BY_NAME = "SELECT t FROM Tag t WHERE LOWER(t.name) LIKE LOWER(:name)";
+    private static final String FIND_ALL = "SELECT t FROM Tag t";
+
+    private static final String FIND_ALL_BY_CERTIFICATE = "SELECT t FROM Tag t LEFT JOIN" +
+            " t.giftCertificates c WHERE c.id = :id";
     @PersistenceContext
     private EntityManager entityManager;
 
     @Override
     public boolean isExists(Tag object) {
-        return findByName(object.getName()).isPresent();
+        boolean result;
+        try {
+            result = findByName(object.getName()).isPresent();
+        } catch (NoResultException e) {
+            return false;
+        }
+        return result;
     }
 
     @Transactional
     @Override
-    public Long save(Tag tag) {
+    public Tag save(Tag tag) {
         entityManager.persist(tag);
-        Long id = tag.getId();
-        log.debug("[TagHibernateRepository.save()] Tag with id:[{}] has been saved.", id);
-        return id;
+        log.debug("[TagHibernateRepository.save()] Tag with id:[{}] has been saved.", tag.getId());
+        return tag;
     }
 
     @Override
     public Optional<Tag> findById(Long id) {
         Tag tag = entityManager.find(Tag.class, id);
-        entityManager.detach(tag);
-        return Optional.of(tag);
+        return Optional.ofNullable(tag);
     }
 
     @Override
     public Optional<Tag> findByName(String name) {
-        TypedQuery<Tag> query = entityManager.createQuery(
-                "SELECT t FROM Tag t WHERE t.name= :name", Tag.class);
+        TypedQuery<Tag> query = entityManager.createQuery(FIND_SINGLE_BY_NAME, Tag.class);
         query.setParameter("name", name);
-        log.debug("save result :{}", query.getSingleResult());
-        return Optional.of(query.getSingleResult());
+        return Optional.ofNullable(query.getSingleResult());
     }
 
     @Override
     public Optional<List<Tag>> findAllByName(String name) {
         TypedQuery<Tag> query = entityManager.createQuery(
-                "SELECT t FROM Tag t WHERE LOWER(t.name) LIKE LOWER(:name)", Tag.class);
+                FIND_ALL_BY_NAME, Tag.class);
         query.setParameter("name", "%" + name + "%");
         return Optional.of(query.getResultList());
     }
 
     @Override
     public Optional<List<Tag>> findAll() {
-        return Optional.of(entityManager.createQuery("SELECT t FROM Tag t", Tag.class)
+        return Optional.of(entityManager.createQuery(FIND_ALL, Tag.class)
                 .getResultList());
     }
 
     @Override
     public Optional<List<Tag>> findAllByCertificate(Long certificateId) {
-        return Optional.empty();
+        TypedQuery<Tag> query = entityManager.createQuery(
+                FIND_ALL_BY_CERTIFICATE, Tag.class);
+        query.setParameter("id",certificateId);
+        return Optional.of(query.getResultList());
     }
 
     @Transactional
     @Override
-    public void deleteById(Long id) {
-        Tag tag = entityManager.find(Tag.class, id);
-        entityManager.remove(tag);
+    public Long deleteById(Long id) {
+        Tag tag;
+        Optional<Tag> tagOptional = findById(id);
+        if (tagOptional.isPresent()) {
+            tag = tagOptional.get();
+            tag.getGiftCertificates().forEach((giftCertificate) -> {
+                giftCertificate.getTags().remove(tag);
+            });
+            log.debug("Tag for removal {}", tag);
+            entityManager.remove(tag);
+            flushAndClear();
+            return tag.getId();
+        }
+        return 0L;
     }
 
+    private void flushAndClear() {
+        entityManager.flush();
+        entityManager.clear();
+    }
 }
